feat: add basic search box
Minimum width of the resizeable parent had to be increased to make space for the editbox.

took the first approach that came to mind with hooking onto `UpdateList` and calling it everytime the text is changed. This should be debounced for performance.

todo: currently the filtered frames are hidden but still take up space in the scrollbox. this needs to be fixed.