major: use `GetLFGDungeonInfo` for localized dungeon names and more
Not all of the expected dungeons can be queried via this API, this is due to limitations of dungeon entries in the classic era `LFGDungeons` db table.

e.g.; "Dire Maul", the SM wings, Naxx, AQ

This api call also has dungeon level ranges so we dont have to manually track those either.